+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Added `query.WithLazyTx(bool)` option for `query.Client.DoTx` calls to enable/disable lazy transactions per operation
+* Added `retry.WithLazyTx(bool)` option for `retry.DoTx` calls to enable/disable lazy transactions for database/sql
+
 ## v3.125.4
 * Added support for `GlobalUniqueIndex` type in `table/options.IndexDescription`